requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
aiofiles==23.2.1
magic-filter==1.0.12
asyncpg==0.29.0
SQLAlchemy==2.0.36
redis==5.0.8
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2025.2
tzlocal==5.3.1
tzdata==2025.2
openai==1.55.3
httpx==0.27.2
qrcode[pil]>=7.4.2
httpcore==1.0.9
anyio==4.11.0
sniffio==1.3.1
requests==2.32.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
mashumaro==3.17
loguru==0.7.3
colorama==0.4.6
win32-setctime==1.2.0
APScheduler==3.10.4
psycopg2-binary==2.9.9